+++ src/Admin/ContactMessageAdmin.php
@@ -156,6 +156,14 @@ final class ContactMessageAdmin extends AbstractBaseAdmin
                         'label' => 'backend.admin.contact.answer',
                     ]
                 )
+                ->add(
+                    'document',
+                    null,
+                    [
+                        'label' => 'backend.admin.contact.attatchment',
+                        'template' => 'Admin/ContactMessage/answer_attatchment_file_preview.html.twig',
+                    ]
+                )
             ;
         }
     }