+++ src/Entity/ContactMessage.php
@@ -3,15 +3,23 @@
 namespace App\Entity;
 
 use App\Entity\Traits\DescriptionTrait;
+use App\Entity\Traits\DocumentFileTrait;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\HttpFoundation\File\File;
+use Symfony\Component\Validator\Constraints as Assert;
+use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
 /**
  * @ORM\Table()
+ *
  * @ORM\Entity(repositoryClass="App\Repository\ContactMessageRepository")
+ *
+ * @Vich\Uploadable
  */
 class ContactMessage extends AbstractBase
 {
     use DescriptionTrait;
+    use DocumentFileTrait;
 
     /**
      * @ORM\Column(type="string", length=255)
@@ -47,6 +55,13 @@ class ContactMessage extends AbstractBase
      * @ORM\Column(type="text", length=4000)
      */
     private ?string $message = null;
+
+    /**
+     * @Vich\UploadableField(mapping="contact_message", fileNameProperty="document")
+     *
+     * @Assert\File(maxSize="10M")
+     */
+    private ?File $documentFile = null;
 
     private bool $privacy;